crop_advisory_backend/requirements.txt
# FastAPI Backend Dependencies for Crop Advisory System
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Firebase Admin SDK for Firestore integration
firebase-admin==6.4.0

# Machine Learning dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2

# External API integration
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1

# Data validation and serialization
pydantic==2.5.1
pydantic-settings==2.1.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-cors==1.7.0

# Caching (optional for API responses)
redis==5.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0